requirements.txt
absl-py==2.1.0
accelerate==0.29.3
adjustText==1.0.4
aiobotocore==2.5.4
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
anndata==0.10.6
annotated-types==0.6.0
annoy==1.17.3
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.5.1
arrow==1.3.0
asciitree==0.3.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autograd==1.6.2
autograd-gamma==0.5.0
axial-positional-embedding==0.2.1
Babel==2.14.0
beautifulsoup4==4.12.3
biopython==1.83
bioservices==1.11.2
bleach==6.1.0
blessed==1.20.0
boltons==23.1.1
botocore==1.31.17
catboost==1.2.5
cattrs==23.2.3
cell2location==0.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.7
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
colorcet==3.1.0
colorlog==6.8.2
CoLT5-attention==0.10.20
comm==0.2.1
contextlib2==21.6.0
contourpy==1.0.5
cospar==0.3.3
ctxcore==0.2.0
cycler==0.11.0
Cython==0.29.37
cytoolz==0.12.3
dask==2024.2.1
dask-image==2023.8.1
datashader==0.16.0
DateTime==5.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.27.2
dill==0.3.8
distributed==2024.2.1
dm-tree==0.1.8
docker-pycreds==0.4.0
docrep==0.3.2
easydev==0.13.1
einops==0.7.0
ERgene==1.2.8
ete3==3.1.3
etils==1.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastcluster==1.2.6
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
fiona==1.9.6
flax==0.8.1
fonttools==4.37.3
formulaic==1.0.1
fqdn==1.5.1
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2023.6.0
future==1.0.0
gdown==5.1.0
geopandas==0.14.3
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
graphtools==1.5.3
graphviz==0.20.3
greenlet==3.0.3
grequests==0.7.0
grpcio==1.62.0
gseapy==0.10.8
h11==0.14.0
h5py==3.7.0
harmonypy==0.0.9
hdbscan==0.8.36
hnswlib==0.8.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
igraph==0.11.4
imageio==2.34.0
importlib-metadata==7.0.1
importlib_resources==6.1.2
inflect==7.0.0
interface-meta==1.3.0
ipykernel==6.29.3
ipython==8.22.2
ipywidgets==8.1.2
isoduration==20.11.0
jax==0.4.25
jaxlib==0.4.25
jedi==0.19.1
Jinja2 @ file:///croot/jinja2_1706733616596/work
jmespath==1.0.1
joblib==1.2.0
json5==0.9.22
jsonlines==4.0.0
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.5.2
jupyterlab==4.1.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keopscore==2.2.3
kiwisolver==1.4.4
ktplotspy==0.2.3
lazy_loader==0.3
leidenalg==0.10.2
lifelines==0.28.0
lightgbm==4.3.0
lightning==2.1.4
lightning-utilities==0.10.1
line-profiler==4.1.2
llvmlite==0.42.0
local-attention==1.9.1
locket==1.0.0
louvain==0.7.1
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
matplotlib==3.8.4
matplotlib-inline==0.1.6
matplotlib-scalebar==0.8.1
mdurl==0.1.2
metatime==1.3.0
mistune==3.0.2
mizani==0.9.3
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml-collections==0.1.1
ml-dtypes==0.3.2
mofax==0.3.6
mpld3 @ file:///croot/mpld3_1696539159604/work
mpmath==1.3.0
msgpack==1.0.8
mudata==0.2.3
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
multiscale_spatial_image==0.11.2
munkres==1.1.4
natsort==8.2.0
nbclient==0.7.4
nbconvert==7.16.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==2.8.6
notebook==7.1.1
notebook_shim==0.2.4
numba==0.59.1
numcodecs==0.12.1
numpy==1.23.4
numpyro==0.14.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.550.52
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
ome-zarr==0.8.3
omegaconf==2.3.0
omicverse==1.5.8
omnipath==1.0.8
opencv-python==4.9.0.80
opt-einsum==3.3.0
optax==0.1.9
orbax-checkpoint==0.5.3
ott==0.1.15
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.1
pandocfilters==1.5.1
param==2.0.2
parso==0.8.3
partd==1.4.1
patsy==0.5.6
pexpect==4.9.0
phate==1.0.11
pillow @ file:///croot/pillow_1707233021655/work
PIMS==0.6.1
platformdirs==4.2.0
plotly==5.22.0
plotnine==0.12.4
POT==0.9.3
prefetch-generator==1.0.3
product-key-memory==0.2.10
progressbar2==4.4.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pybedtools==0.9.1
pybind11==2.11.1
pycparser==2.21
pyct==0.5.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeseq2==0.4.7
pyg-lib==0.4.0+pt22cu121
pygam==0.8.0
pygeos==0.14
Pygments==2.17.2
PyGSP==0.5.1
pykeops==2.2.3
pynndescent==0.5.7
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
pyproj==3.6.1
pyro-api==0.1.2
pyro-ppl==1.9.0
pysam==0.22.0
PySocks==1.7.1
python-circos==0.3.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotplot==0.0.1
python-json-logger==2.0.7
python-utils==3.8.2
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
reformer-pytorch==1.4.4
regex==2024.4.28
requests==2.31.0
requests-cache==1.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
s-gd2==1.8.1
s3fs==2023.6.0
safetensors==0.4.3
scanpy==1.9.8
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scikit-misc==0.3.1
scipy==1.12.0
scprep==1.2.3
scrublet==0.2.3
scvi-tools==1.1.2
seaborn==0.13.2
Send2Trash==1.8.2
sentry-sdk==1.45.0
session-info==1.0.0
setproctitle==1.3.3
shapely==2.0.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicerator==1.1.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spatial_image==0.3.0
spatialdata==0.0.16
squidpy==1.4.1
stack-data==0.6.3
statsmodels==0.14.1
stdlib-list==0.8.0
suds-community==1.1.2
sympy==1.12
tabulate==0.9.0
tangram-sc==1.0.4
tasklogger==1.2.0
tblib==3.0.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorstore==0.1.54
termcolor==2.4.0
terminado==0.18.0
texttable==1.6.4
threadpoolctl==3.1.0
tifffile==2024.2.12
timm==0.9.16
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
torch-sparse==0.6.18
torch_cluster==1.6.3+pt22cu121
torch_geometric==2.5.0
torch_scatter==2.1.2+pt22cu121
torch_spline_conv==1.2.2+pt22cu121
torchaudio==2.2.1
torchmetrics==1.3.1
torchsummary==1.5.1
torchvision==0.17.1
tornado==6.4
tqdm==4.64.1
traitlets==5.14.1
transformers==4.40.1
triton==2.2.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.3
UpSetPlot==0.9.0
uri-template==1.3.0
url-normalize==1.4.3
urllib3==1.26.18
validators==0.23.2
voila==0.5.5
wandb==0.16.6
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
xarray==2023.12.0
xarray-dataclasses==1.7.0
xarray-datatree==0.0.14
xarray-schema==0.0.3
xarray-spatial==0.3.7
xgboost==2.0.3
xmltodict==0.13.0
yarl==1.9.4
zarr==2.17.1
zict==3.0.0
zipp==3.17.0
zope.event==5.0
zope.interface==6.2
